third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/history-traversal/browsing_context_name-expected.txt

This is a testharness.js-based test.
FAIL Retaining window.name on history traversalassert_equals: expected 2 but got 1
Harness: the test ran to completion.
